cheri: Manually apply fixes for FixPtrToULongCast warnings
Manually add fixes for the FixPtrToULongCast warnings where
clang-tidy isn't able to fix them automatically due to macro
expansions.

Explicitly ignore warnings in fs/btrfs/ because this file system
plays tricks with pointers that need a different fix.

Ignore some warnings that need a different fix.

+++ include/linux/rhashtable.h
@@ -101,7 +101,7 @@ struct bucket_table {
  * Instead we store a NULL
  */
 #define	RHT_NULLS_MARKER(ptr)	\
-	((void *)NULLS_MARKER(((unsigned long) (ptr)) >> 1))
+	((void *)NULLS_MARKER(((uintptr_t) (ptr)) >> 1))
 #define INIT_RHT_NULLS_HEAD(ptr)	\
 	((ptr) = NULL)
